Created a new directory to organize kotlin files
