+++ app/src/main/java/com/example/newleaf2022/AccountsFragment.kt
@@ -7,10 +7,10 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.activityViewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.example.newleaf2022.adapters.AccountsAdapter
-import com.example.newleaf2022.adapters.CategoryAdapter
 import com.example.newleaf2022.databinding.FragmentAccountsBinding
+import com.example.newleaf2022.dataclasses.Accounts
+import com.example.newleaf2022.dataclasses.Budgets
 import com.example.newleaf2022.viewmodels.*
 
 class AccountsFragment : Fragment() {
