+++ app/src/main/java/com/example/newleaf2022/dataclasses/Accounts.kt
@@ -1,4 +1,4 @@
-package com.example.newleaf2022.viewmodels
+package com.example.newleaf2022.dataclasses
 
 data class Accounts(val name: String,
                     val type: String,
