+++ app/src/main/java/com/example/newleaf2022/dataclasses/Categories.kt
@@ -1,4 +1,4 @@
-package com.example.newleaf2022.viewmodels
+package com.example.newleaf2022.dataclasses
 
 data class Categories(val name: String,
                       val subcategories: ArrayList<Subcategories>) {
