+++ app/src/main/java/com/example/newleaf2022/dataclasses/Subcategories.kt
@@ -1,3 +1,3 @@
-package com.example.newleaf2022.viewmodels
+package com.example.newleaf2022.dataclasses
 
 data class Subcategories(val name: String)
